fix bug

+++ core/data/batch_creator.py
@@ -55,7 +55,7 @@ class BatchCreator(data.Dataset):
         """解析数据"""
         for input_data in input:
             with h5py.File(input_data, 'r') as raw:
-                self.input_data.append((raw['image'].value-128) / 33.0)
+                self.input_data.append((raw['image'].value.astype(np.float32)-128) / 33.0)
                 self.label_data.append(raw['label'].value)
                 self.coor.append(raw['coor'].value)
                 self.seed.append(logit(np.full(list(raw['image'].value.shape), 0.05, dtype=np.float32)))
